Split out storage-specific parts of Index struct
